Refactored existing code, still require calculating relative rotation of thruster outputs.
Simplified main quadcopter class, the main functionality of the quadcopter assumes that the thrusters are always initially set at a rotation away from the origin - not relatively rotated based on the orientation of the quadcopter. This relative rotation will require that there is an additional degree of freedom in each of the thrusters.