Proper rotation about Y axis, fixed basic assumptions for higher accuracy simulation
Coordinates and thrusters are now dynamically modified to account for the change in rotation - when the system rotates each of the individual feedback controllers need to understand this difference to be able to adjust its output to reach its target. This setup requires the usage of the ADRC or the difficult to calibrate adaptive PID. The effect of this change can only be seen about the Y axis, the X and Z axes respond inappropriately still - causing the system to have a major potential to move as opposed to just rotating. This cannot be tested through removing gravity as no negative thruster output can be attained and the current calculation method cannot allow rotation more than 90 degrees.

The next commit should have the working X and Z rotational axes, not while causing the system to have a potential to move.