Partial rotation for negative values of Y in direction vector
Rotation of each joint independently on their own axes responds appropriately. The directionality is flipped however when approaching the axis. Responds as a point coordinate on a sphere where the sphere is split in half and the bottom hemisphere has 2 vertical slices with even separation where two of the opposing sides are pulled and rotated outward while the other two are kept inward - as a normal sphere face. As the rotation approaches another axis it flips which bottom slices of the hemisphere are pulled outward.

This function still requires work, but demonstrates each independent full rotation appropriately