Semi-continuous function for calculating hover angles
Given any direction-angle coordinate, hover angles can be calculated. The calculation, based on atan2 and asin are limited due to the discontinuity in primarily the atan2 function where the value can switch from previously being -179 degrees to currently being 180 degrees; this causes a rapid change on the primary joint of the quadcopter when this limitation is reached. This can be corrected through taking into account the previous angle, determining what the shortest route is; for instance, switching from -179 degrees to -180 degrees. This essentially removes the pi to -pi limitation of the atan2 function.