Added test interface for C++/CLI  and C++ interface
Switching from DLL to EXE compilation in the C++/CLI application results in successful testing of the C++ code without running into a dereferencing null pointer exception. It can be concluded from this information, and the relevant stack trace that the C# code is generating the exception. The C# code may be doing this on its own or through the direct referencing of either a DTRQCS object or SQuad object. More testing is necessary to determine the direct source of the problem or the debugging needs to be configured to allow for proper stepping through the linked projects.